Remove cyclic import This commit moves per-card validation from tabs/config.py to config_io.py to remove a cyclic import between those two files.
